requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
python-multipart

sentence-transformers
transformers
torch
faiss-cpu
scikit-learn
pandas
numpy
scipy

huggingface_hub
safetensors

SpeechRecognition
edge-tts
pygame
pvporcupine
PyAudio

openpyxl
rich
typer
groq